refactor(plugins):fix typing errpr
