+++ packages/pc-plugins/src/fisheye/index.ts
@@ -1,6 +1,5 @@
 import { clone } from '@antv/util';
-import { IG6GraphEvent, ShapeStyle } from '@antv/g6-core/es/types';
-import { Graph } from '@antv/g6';
+import { IG6GraphEvent, ShapeStyle, Graph } from '@antv/g6';
 import Base from '../base';
 
 const DELTA = 0.05;
