+++ packages/pc-plugins/src/menu/index.ts
@@ -1,8 +1,7 @@
 import { modifyCSS, createDom } from '@antv/dom-util';
 import { isString } from '@antv/util';
 import insertCss from 'insert-css';
-import { Graph } from '@antv/g6';
-import { IG6GraphEvent, Item } from '@antv/g6-core/es/types';
+import { IGraph, IG6GraphEvent, Item } from '@antv/g6';
 import Base, { IPluginBaseConfig } from '../base';
 
 insertCss(`
@@ -104,7 +103,7 @@ export default class Menu extends Base {
 
     const menuDom = this.get('menu');
     const getContent = this.get('getContent');
-    const graph: Graph = this.get('graph');
+    const graph: IGraph = this.get('graph');
     const menu = getContent(e, graph);
     if (isString(menu)) {
       menuDom.innerHTML = menu;
