+++ packages/pc-plugins/src/timeBar/index.ts
@@ -8,9 +8,8 @@ import { createDom, modifyCSS } from '@antv/dom-util';
 import Base, { IPluginBaseConfig } from '../base';
 import TrendTimeBar, { SliderOption } from './trendTimeBar';
 import TimeBarSlice, { TimeBarSliceOption } from './timeBarSlice';
-import { IGraph } from '@antv/g6/es/interface/graph';
 import { VALUE_CHANGE } from './constant';
-import { GraphData, IG6GraphEvent, ShapeStyle, TimeBarType } from '@antv/g6-core';
+import { GraphData, IG6GraphEvent, ShapeStyle, TimeBarType, IGraph } from '@antv/g6';
 import { Interval } from './trend';
 import { ControllerCfg } from './controllerBtn';
 import { isString } from '@antv/util';
