+++ packages/pc-plugins/src/timeBar/timeButton.ts
@@ -4,7 +4,7 @@
 
 import { IGroup, IShape } from '@antv/g-base';
 import { deepMix } from '@antv/util';
-import { ShapeStyle } from '@antv/g6-core/es/types';
+import { ShapeStyle } from '@antv/g6';
 
 /** 播放按钮配置 */
 interface ButtonCfg {
