+++ packages/pc-plugins/src/timeBar/trendTimeBar.ts
@@ -3,8 +3,8 @@ import { get, size, assign, each } from '@antv/util';
 import Trend, { TrendCfg } from './trend';
 import Handler from './handler';
 import ControllerBtn, { ControllerCfg } from './controllerBtn';
-import { IGraph } from '@antv/g6/es/interface/graph';
-import { ShapeStyle } from '@antv/g6-core/es/types';
+
+import { ShapeStyle, IGraph } from '@antv/g6';
 import {
   VALUE_CHANGE,
   TIMELINE_START,
