+++ packages/pc-plugins/src/tooltip/index.ts
@@ -1,10 +1,8 @@
 import { modifyCSS, createDom } from '@antv/dom-util';
 import { isString } from '@antv/util';
 import insertCss from 'insert-css';
-import { Graph } from '@antv/g6';
-import { IG6GraphEvent, Item } from '@antv/g6-core/es/types';
+import { Graph, IG6GraphEvent, Item, IGraph } from '@antv/g6';
 import Base, { IPluginBaseConfig } from '../base';
-import { IGraph } from '@antv/g6/es/interface/graph';
 
 insertCss(`
   .g6-component-tooltip {
