+++ packages/pc-plugins/tests/unit/menu-spec.ts
@@ -1,5 +1,5 @@
 import G6 from '@antv/g6';
-import { G6GraphEvent, IG6GraphEvent } from '@antv/g6-core';
+import { G6GraphEvent, IG6GraphEvent } from '@antv/g6';
 import Menu from '../../src/menu';
 
 const div = document.createElement('div');
